Allow spans to be sampled client-side (head-based sampling)
This functionality is exposed in the otel SDK, but not the lightstep
launcher. Per this blog post:

https://www.aspecto.io/blog/opentelemetry-sampling-everything-you-need-to-know/

Sampling client-side is also called "head-based sampling", while doing
it at the microsatellite (either as a fixed percent across all services
or with a configurable percent per-service) is tail-based sampling.

I'd say there's a reasonable argument to support both head and tail
based sampling in the lightstep launcher.

Signed-off-by: Nick Thomas <nthomas@kentik.com>