Add span and metric instrumentation to the LS metrics SDK export (#590)
**Description:** Adds a span and a metric about the outcome of the LS
metrics SDK periodic exporter. Changes the default to enable spans about
the metrics SDK export. We will consider doing this for the span
exporter, but it's different in that case because of the processor.
Metrics SDKs don't drop, so we keep it simple in this PR w/ only metrics
support.

**Link to tracking Issue:** Internal: LS-56746

**Testing:** Manual testing. Also: `go work sync` because otel deps were
out of date w/ respect to OTel-Arrow.

**Documentation:** None. Self-tracing/metricing is not documented,
happens as a result of the otelcol export pipeline using the globals.