exporter: allow custom meter/trace providers (#739)
This adds a feature to allow passing in a custom trace and metric
provider for the self-reporting features.

To do this, alter `NewExporter` to employ the variadic options pattern
over a new options struct, `ExporterOptions`. `Options` is already used
by the some `Config` building contraptions.

This does make the package namespace a little crowded... `With*`
functions can apply to either the exporter or the config, but I don't
know that makes it confusing enough to split things up yet.