add prometheus exporter (#775)
**Description:**

I've added a prometheus exporter, following the same pattern that was
used for the OTel exporter. This is my first time working with this
code-base, so I've left lots of questions as GitHub comments!

**Testing:** 

I wrote an end-to-end test that makes sure that `/metrics` endpoint is
serving the correct data. I'm going to add more tests once @jmacd signs
off that the high-level direction of the PR looks good.

---------

Co-authored-by: Isaac Zinda <isaac@lightstep.com>