Self telemetry tidy-up (#798)
**Description:** There was an accidental regression in self-telemetry
setup. This was discovered while updating to the latest release. Fixed
here, plus more idiomatic Go.

Note that the Trace exporter setup allows options to override the
providers, which is used for internal test purposes. The Metric exporter
does not have the same customization, so the Prom/OTLP metric exporters
are somewhat simpler.